update footer files
Przeniesienie kodu z main.js do footer.js i podpięcie do index.js i library.js

+dodanie zamknięcie okna dialogowego w footer na kliknięcie poza modalem
