+++ src/js/main.js
@@ -8,26 +8,6 @@ import { moviesLoading } from './services/loader';
 import { reloadHeader, setRegisterAndSignUp } from './services/firebase';
 
 /* --------------------------------- FOOTER --------------------------------- */
-const dialog = document.getElementById('footer-dialog');
-const hideDialogBtn = document.getElementById('hide');
-const showDialogBtn = document.getElementById('show');
-const body = document.querySelector('body');
-
-reloadHeader();
-setRegisterAndSignUp();
-
-showDialogBtn.addEventListener('click', () => {
-  body.style.overflow = 'hidden';
-  dialog.showModal();
-});
-
-const closeModal = () => {
-  body.style.overflow = 'auto';
-  dialog.close();
-};
-
-hideDialogBtn.addEventListener('click', closeModal);
-dialog.addEventListener('close', closeModal);
 
 /* --------------------------------- MOVIE-MODAL --------------------------------- */
