+++ src/library.js
@@ -3,6 +3,7 @@ import './js/services/detailModal';
 import { reloadHeader, setRegisterAndSignUp } from './js/services/firebase';
 import { initializeLibrary } from './js/services/loadMovies';
 import { moviesLoading } from './js/services/loader';
+import './js/services/footer';
 
 reloadHeader();
 setRegisterAndSignUp();
